refactor tribe member card
